Refactor: Rename modules, optimize iterator and sync primitives
